[core][configuration] Move high level conf mgmt to configuration package
