+++ core/server.go
@@ -35,9 +35,9 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/AliceO2Group/Control/common/product"
+	"github.com/AliceO2Group/Control/configuration/the"
 	"github.com/AliceO2Group/Control/core/task"
 	"github.com/AliceO2Group/Control/core/task/channel"
-	"github.com/AliceO2Group/Control/core/the"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
