+++ core/task/task.go
@@ -45,9 +45,9 @@ import (
 	"github.com/AliceO2Group/Control/common/utils"
 	"github.com/AliceO2Group/Control/common/event"
 	"github.com/AliceO2Group/Control/common/utils/uid"
+	"github.com/AliceO2Group/Control/configuration/template"
 	"github.com/AliceO2Group/Control/core/controlcommands"
 	"github.com/AliceO2Group/Control/core/task/channel"
-	"github.com/AliceO2Group/Control/core/workflow/template"
 	"github.com/mesos/mesos-go/api/v1/lib"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
