+++ core/workflow/aggregatorrole.go
@@ -30,9 +30,9 @@ import (
 	texttemplate "text/template"
 
 	"github.com/AliceO2Group/Control/common/event"
-	"github.com/AliceO2Group/Control/core/repos"
+	"github.com/AliceO2Group/Control/configuration/repos"
+	"github.com/AliceO2Group/Control/configuration/template"
 	"github.com/AliceO2Group/Control/core/task"
-	"github.com/AliceO2Group/Control/core/workflow/template"
 	"github.com/gobwas/glob"
 	"github.com/sirupsen/logrus"
 )
