+++ core/workflow/taskrole.go
@@ -31,9 +31,9 @@ import (
 	"time"
 
 	"github.com/AliceO2Group/Control/common/event"
-	"github.com/AliceO2Group/Control/core/repos"
+	"github.com/AliceO2Group/Control/configuration/repos"
+	"github.com/AliceO2Group/Control/configuration/template"
 	"github.com/AliceO2Group/Control/core/task"
-	"github.com/AliceO2Group/Control/core/workflow/template"
 	"github.com/gobwas/glob"
 )
